fix: big hack/workaround to support Safari/iOS
Safari has many problems and limitations, some features needs to be a
reply of `click` event. However, in the case of `window.Open` that isn't
enough.

The alternative is to create a big `<a>` tag, which covers all the page,
including Gio itself. So, the next click will hit the `<a>` instead of
the `<canvas>`. Then, the link will open.

The `<a>` will be removed as a response to `system.StageEvent`, which uses
the `Page Visibility API`. So, when you click it loses the focus and then we
remove the `<a>`.

That is a big hack, but I don't see any other way to fix.